requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
matplotlib>=3.7.0
flask>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0
scikit-learn>=1.3.0
python-telegram-bot>=20.0
